+++ src/main/java/com/mobble/mobbleserver/application/chat/room/service/club/ClubChatRoomQueryService.java
@@ -9,7 +9,7 @@ import com.mobble.mobbleserver.domain.chat.room.Participant;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
-import com.mobble.mobbleserver.refactor.chat.chatMessage.repository.ChatMessageRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.chat.message.ChatMessageRepository;
 import com.mobble.mobbleserver.infrastructure.web.chat.room.club.dto.response.ClubChatRoomPreviewResponseDto;
 import com.mobble.mobbleserver.refactor.club.core.entity.Club;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
