+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/chat/message/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.chat.chatMessage.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.chat.message;
 
 import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
