+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/chat/room/common/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.infrastructure.persistence.chat.common;
+package com.mobble.mobbleserver.infrastructure.persistence.chat.room.common;
 
 import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
