+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/chat/room/direct/DirectRoomInfoQueryDslRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.infrastructure.persistence.chat.direct;
+package com.mobble.mobbleserver.infrastructure.persistence.chat.room.direct;
 
 import com.mobble.mobbleserver.domain.chat.room.DirectRoomInfo;
 import com.querydsl.jpa.impl.JPAQueryFactory;
