+++ src/main/java/com/mobble/mobbleserver/refactor/chat/chatMessage/validator/ChatMessageValidator.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.chat.chatMessage.validator;
 
 import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
-import com.mobble.mobbleserver.refactor.chat.chatMessage.repository.ChatMessageRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.chat.message.ChatMessageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
