+++ src/main/java/com/mobble/mobbleserver/refactor/chat/directChatRoom/service/DirectChatRoomService.java
@@ -5,8 +5,8 @@ import com.mobble.mobbleserver.domain.chat.message.ChatMessage;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
-import com.mobble.mobbleserver.refactor.chat.chatMessage.dto.request.ChatMessageRequestDto;
-import com.mobble.mobbleserver.refactor.chat.chatMessage.dto.response.ChatMessageResponseDto;
+import com.mobble.mobbleserver.infrastructure.web.chat.message.dto.request.ChatMessageRequestDto;
+import com.mobble.mobbleserver.infrastructure.web.chat.message.dto.response.ChatMessageResponseDto;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.service.ChatMessageService;
 import com.mobble.mobbleserver.infrastructure.web.chat.room.direct.dto.request.DirectChatMessageRequestDto;
 import com.mobble.mobbleserver.infrastructure.web.chat.room.direct.dto.response.DirectChatMessageResponseDto;
